+++ src/components/style/VcsFillMenu.vue
@@ -1,0 +1,56 @@
+<template>
+  <MenuWrapper
+    v-bind="{ value, valueDefault }"
+    :value-fallback="{ color: [255, 255, 255, 1] }"
+    v-on="$listeners"
+    name="components.style.fill"
+  >
+    <template #preview>
+      <v-sheet :color="rgbaString" width="100%" height="100%" />
+    </template>
+    <template #content>
+      <VcsFillSelector :value="value" v-on="$listeners" />
+    </template>
+  </MenuWrapper>
+</template>
+
+<script>
+  import { computed } from 'vue';
+  import { VSheet } from 'vuetify/lib';
+  import { VcsFillSelector } from '@vcmap/ui';
+  import MenuWrapper from './MenuWrapper.vue';
+  import { useColorObject, rgbaObjectToString } from './composables.js';
+
+  /**
+   * @description A wrapper for the VcsFillSelector, that has a small color preview and a menu that pops up when clicking the preview, containing the fill selector.
+   * When clicking the reset button, the valueDefault is emitted, when unchecking the checkbox in front of the preview, null is emitted. If it is checked again, valueDefault is emitted. If the valueDefault is undefined or null, { color: [255, 255, 255, 1] } is emitted.
+   * @vue-prop {import("ol/style/Fill").Options} [value] - The Fill Options
+   * @vue-prop {import("ol/style/Fill").Options} [valueDefault] - The default Fill Options
+   */
+  export default {
+    name: 'VcsFillMenu',
+    components: {
+      VSheet,
+      VcsFillSelector,
+      MenuWrapper,
+    },
+    props: {
+      value: {
+        type: Object,
+        default: undefined,
+      },
+      valueDefault: {
+        type: Object,
+        default: undefined,
+      },
+    },
+    setup(props) {
+      const rgbaObject = useColorObject(() => props.value?.color);
+      return {
+        rgbaString: computed(() => rgbaObjectToString(rgbaObject.value)),
+      };
+    },
+  };
+</script>
+
+<style scoped></style>